VLC Tx finished

+++ VLC_TX_GD32/VLC_Drivers/Inc/VLC_timer.h
@@ -12,7 +12,7 @@
  * Using APB Timer to delay, in us.
  * @param us The delay time.
  */
-void VLC_udelay(uint16_t us);
+void VLC_udelay(uint32_t us);
 
 
 #endif //VLC_TX_GD32_VLC_TIMER_H
