+++ VLC_TX_GD32/VLC_Drivers/Src/VLC_timer.c
@@ -2,12 +2,12 @@
 // Created by Administrator on 11/30/2023.
 //
 
-#include "VLC_timer.h"
+#include "../Inc/VLC_timer.h"
 
 #include "tim.h"
 
 
-void VLC_udelay(uint16_t us) {
+void VLC_udelay(uint32_t us) {
     __HAL_TIM_SET_COUNTER(&htim2, 0);
     while ((__HAL_TIM_GET_COUNTER(&htim2)) < us);
 }
